extract component Ranking
